+++ tests/test_gitlab_checkout_mrs.py
@@ -121,7 +121,7 @@ class TestGitLabCheckoutMRs(BaseTest):
             |
             o-hotfix/add-trigger
               |
-              o-ignore-trailing  MR !3 (some_other_user) rebase=no push=no
+              o-ignore-trailing
                 |
                 o-chore/fields
 
@@ -147,7 +147,7 @@ class TestGitLabCheckoutMRs(BaseTest):
             |
             o-hotfix/add-trigger
               |
-              o-ignore-trailing  MR !3 (some_other_user) rebase=no push=no
+              o-ignore-trailing
                 |
                 o-chore/fields
 
@@ -180,7 +180,7 @@ class TestGitLabCheckoutMRs(BaseTest):
             |
             o-hotfix/add-trigger
               |
-              o-ignore-trailing  MR !3 (some_other_user) rebase=no push=no
+              o-ignore-trailing
                 |
                 o-chore/fields